CI: limit od to 1gib for _ddt
While wiring up fstests into github actions I noticed that it would hang
on one of our tests that uses _ddt.  A long investigation uncovered that
od was still reading /dev/urandom and spewing to stdout despite the fact
that it had been disconnected.  This is a bug in od, however it's going
to be a while until the fix makes it everywhere, so for now change the
_ddt helper to take an argument for od to limit the amount it reads.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>