btrfs: test scan but not register the single device fs
Recently, in the kernel commit 0d9436739af2 ("btrfs: scan but don't
register device on single device filesystem"), we adopted an approach
where we scan the device to validate it. However, we do not register
it in the kernel memory since it is not required to be remembered.

However, the seed device should continue to be registered because
otherwise, the mount operation for the sprout device will fail.

This patch ensures that we honor the mount requirements and do not break
anything while making changes in this part of the code.

Signed-off-by: Anand Jain <anand.jain@oracle.com>
Reviewed-by: Zorro Lang <zlang@redhat.com>
Signed-off-by: Zorro Lang <zlang@kernel.org>