generic/251: check min and max length and minlen for FSTRIM
Every now and then, this test fails with the following output when
running against my development tree when configured with an 8k fs block
size:

  --- a/tests/generic/251.out	2023-07-11 12:18:21.624971186 -0700
  +++ b/tests/generic/251.out.bad	2023-10-15 20:54:44.636000000 -0700
  @@ -1,2 +1,4677 @@
   QA output created by 251
   Running the test: done.
  +fstrim: /opt: FITRIM ioctl failed: Invalid argument
  +fstrim: /opt: FITRIM ioctl failed: Invalid argument
  ...
  +fstrim: /opt: FITRIM ioctl failed: Invalid argument

Dumping the exact fstrim command lines to seqres.full produces this at
the end:

  /usr/sbin/fstrim -m 32544k -o 30247k -l 4k /opt
  /usr/sbin/fstrim -m 32544k -o 30251k -l 4k /opt
  ...
  /usr/sbin/fstrim -m 32544k -o 30255k -l 4k /opt

The count of failure messages is the same as the count as the "-l 4k"
fstrim invocations.  Since this is an 8k-block filesystem, the -l
parameter is clearly incorrect.  The test computes random -m and -l
options.

Therefore, create helper functions to guess at the minimum and maximum
length and minlen parameters that can be used with the fstrim program.
In the inner loop of the test, make sure that our choices for -m and -l
fall within those constraints.

Signed-off-by: "Darrick J. Wong" <djwong@kernel.org>
Reviewed-by: Zorro Lang <zlang@redhat.com>
Signed-off-by: Zorro Lang <zlang@kernel.org>